+++ src/test/resources/inspections/MapFlattenToFlatMapInspection/map_flatten-to-flatmap-with-indentation.after.php
@@ -1,0 +1,6 @@
+<?php
+
+collect($shopifyOrder->fulfillments)
+    ->flatMap(
+        fn (ShopifyFulfillment $fulfillment) => $this->mapLineItems($order, $shopifyOrder, $fulfillment)
+    );